requirements.txt
python-telegram-bot>=21.0
openai>=1.50.0
litellm>=1.40.0
python-dotenv>=1.0.0
sqlalchemy[asyncio]>=2.0.0
asyncpg>=0.29.0
pydantic>=2.5.0
pydantic-settings>=2.1.0
apscheduler>=3.10.4
structlog>=24.1.0
tenacity>=8.2.3
pytz>=2025.2
dateparser>=1.3.0
watchdog>=6.0.0
streamlit>=1.40.0
psycopg2-binary>=2.9.0
cachetools>=5.3.0
fastapi>=0.110.0
uvicorn>=0.27.0
python-multipart>=0.0.9
spotipy>=2.23.0
google-api-python-client>=2.120.0
google-auth-oauthlib>=1.2.0
google-auth-httplib2>=0.2.0
google-auth>=2.28.0
